Move test setup to TestMain
