+++ things/api/grpc/setup_test.go
@@ -1,0 +1,16 @@
+
+package grpc_test
+
+import (
+	"os"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	svc = newService(map[string]string{token: email})
+	startGRPCServer(svc, port)
+
+	code := m.Run()
+
+	os.Exit(code)
+}
